Teleporting through blocks fixed. Collision slightly different
Problem was originating from the fact that the y-velocity
was never being reset to 0 when hitting the ground.
Therefore, due to gravity acceleration every frame,
eventually the y-velocity would be so high that the player
would clip through the block.
y-velocity is now set to 0 whenever a collision delta is sent
to the player that has a y part greater than 0.