Collision between blocks and player works, with caveat
Most directions of collision seem to be working, however
when a player is shoved up against the side of a wall, an
interesting effect comes into play where the player will
slowly slide down the wall, jumping only moves them a
little, and they can jump infinitely while against the wall.

Not sure what the cause is yet.